fix: added create new version for toolset and app assets
