--- apps/ai-dial-admin/src/components/Assets/Prompts/utils.ts
@@ -1,11 +1,0 @@
-import { DialPrompt } from '@/src/models/dial/prompt';
-import { modifyNameVersionInPrompt } from '@/src/utils/prompts/versions';
-
-export const addNewVersion = (entity: DialPrompt, version: string) => {
-  const path = modifyNameVersionInPrompt(entity.path, void 0, version);
-  return {
-    ...entity,
-    path,
-    version,
-  };
-};
