+++ apps/ai-dial-admin/src/components/Assets/tests/utils.spec.ts
@@ -1,6 +1,6 @@
 import { DialPrompt } from '@/src/models/dial/prompt';
 import { describe, expect, test } from 'vitest';
-import { filterLatestVersions, getEntityForUpdate, getIsNeedToMove, getVersionsPerName } from '../utils';
+import { addNewVersion, filterLatestVersions, getEntityForUpdate, getIsNeedToMove, getVersionsPerName } from '../utils';
 
 describe('filterLatestVersions', () => {
   test('Should return only latest versions', () => {
@@ -74,5 +74,29 @@ describe('getEntityForUpdate', () => {
     const result = getEntityForUpdate(entity, undefined);
     expect(result.folderId).toBeUndefined();
     expect(result.name).toBe('Prompt');
+  });
+});
+
+describe('addNewVersion', () => {
+  test('should correctly change version and path if version is numeric', () => {
+    const entity = { folderId: '2', name: 'Prompt', path: 'somePath__0.0.1' } as any;
+    const result = addNewVersion(entity, '1.2.3');
+    expect(result).toEqual({
+      folderId: '2',
+      name: 'Prompt',
+      path: 'somePath__1.2.3',
+      version: '1.2.3',
+    });
+  });
+
+  test('should correctly change version and path if version is not numeric', () => {
+    const entity = { folderId: '2', name: 'Prompt', path: 'somePath__oldVersion' } as any;
+    const result = addNewVersion(entity, 'newVersion');
+    expect(result).toEqual({
+      folderId: '2',
+      name: 'Prompt',
+      path: 'somePath__newVersion',
+      version: 'newVersion',
+    });
   });
 });
