+++ apps/ai-dial-admin/src/components/Assets/utils.ts
@@ -1,6 +1,6 @@
 import { Asset } from '@/src/models/dial/deployment-asset';
 import { DialPrompt } from '@/src/models/dial/prompt';
-import { compareVersions } from '@/src/utils/prompts/versions';
+import { compareVersions, modifyNameVersionInPrompt } from '@/src/utils/prompts/versions';
 
 export const filterLatestVersions = (data: Asset[]) => {
   const latestVersions: Record<string, DialPrompt> = {};
@@ -42,5 +42,14 @@ export const getEntityForUpdate = (entity: Asset, initialEntity?: Asset) => {
   return {
     ...entity,
     folderId: (initialEntity as Asset)?.folderId,
+  };
+};
+
+export const addNewVersion = (entity: Asset, version: string) => {
+  const path = modifyNameVersionInPrompt(entity.path, void 0, version);
+  return {
+    ...entity,
+    path,
+    version,
   };
 };
